src/chat/requirements.txt
Flask
flask-cors
langchain
langchain-community
langchain-openai
openai
faiss-cpu
pypdf
python-dotenv